Factoring out static function
