+++ pisa_utils/analyse.py
@@ -1,10 +1,8 @@
 import json
 import logging
 import os
-import xml.etree.ElementTree as ET
 from gemmi import cif
-
-from pisa_utils.run_pisa import run_pisalite
+from pisa_utils.utils import parse_xml_file
 
 logger = logging.getLogger()
 
@@ -27,21 +25,21 @@ class AnalysePisa:
 
     def get_pisa_result(self, input_file, interfaces_xml_file, assembly_xml_file,pdb_id,pisa_config):
         """
-        This function runs pisa-lite to obtain interface and xml files 
+        This function runs pisa-lite to obtain interface and xml files
 
         @params:
           input_file: type str - path to assembly configuration cif file
-          interfaces_xml_file : type str -  file paths to  interfaces.xml and assembly.xml 
-          assembly_xml_file: type str -  file paths to assembly.xml 
-          pdb_id : type str - entry pdb 
-          pisa_config: type str path to pisa configuration file 
+          interfaces_xml_file : type str -  file paths to  interfaces.xml and assembly.xml
+          assembly_xml_file: type str -  file paths to assembly.xml
+          pdb_id : type str - entry pdb
+          pisa_config: type str path to pisa configuration file
 
         Returns : None
 
         Raises:
         Failing to run Pisa-Lite error
 
-        Requirements: 
+        Requirements:
         - This function runs pisa-lite code assuming binary self.pisa_binary is already provided
         """
 
@@ -71,13 +69,13 @@ class AnalysePisa:
         unp_acc=None
         unp_num=None
 
-        
+
         path=os.path.join(self.input_updated_cif,'{}_updated.cif'.format(self.pdb_id)) #updated cif file path
 
         #Reading uniprot acc and seq numbers in updated cif file:
 
 
-        doc = cif.read_file(path)  
+        doc = cif.read_file(path)
         block = doc.sole_block()
         label_seq_id=block.find_loop("_atom_site.label_seq_id")
         atom_name=block.find_loop("_atom_site.label_atom_id")
@@ -87,7 +85,7 @@ class AnalysePisa:
         db_num= block.find_loop("_atom_site.pdbx_sifts_xref_db_num")
         res_name=block.find_loop("_atom_site.label_comp_id")
 
-        
+
         n=0 #counts if atom is not found in updated cif file ****
 
         if (int(int_lab_seqnum)<0): int_lab_seqnum="." # if sequence identifier read in pisa-lite is not available, replace with a dot
@@ -100,32 +98,19 @@ class AnalysePisa:
                 if (dbname=="UNP"):
                     unp_acc=dbacc
                     unp_num=dbnum
-                    
+
                     return unp_acc,unp_num
                 else:
                     logging.debug('No UNP numbers found for atom: name {},label_seq_id {},seq_num, residue {}'.format(int_atname,int_lab_seqnum,int_seqnum,int_resname))
                     unp_acc=None
                     unp_num=None
                     return unp_acc,unp_num
-        if n==0: # If atom was not found in updated cif file, return message 
-            
+        if n==0: # If atom was not found in updated cif file, return message
+
             logging.debug('atom not found in updated cif file: name {},label_seq_id {},seq_num, residue {}'.format\
 (int_atname,int_lab_seqnum,int_seqnum,int_resname))
-                
+
             return unp_acc,unp_num
-                
-    @staticmethod
-    def parse_xml_file(xml_file):
-        
-        root = None
-        try:
-            logging.debug('parsing: {}'.format(xml_file))
-            tree = ET.parse(xml_file)
-            root = tree.getroot()
-        except Exception as e:
-            logging.error('invalid xml file: {}'.format(xml_file))
-            logging.error(e)
-        return root
 
     def get_bond_dictionary(self,bonds,bondtype):
 
@@ -142,7 +127,7 @@ class AnalysePisa:
         bond_dict - Bonds dictionary 
  
         """
-        
+
         atom_site1_chains=[]
         atom_site1_residues=[]
         atom_site1_label_asym_ids=[]
@@ -153,14 +138,14 @@ class AnalysePisa:
         atom_site1_label_seq_ids=[]
         atom_site1_label_atom_ids=[]
         atom_site1_inscodes=[]
-        
+
         atom_site2_chains=[]
         atom_site2_residues=[]
         atom_site2_label_asym_ids=[]
         atom_site2_orig_label_asym_ids=[]
         atom_site2_unp_nums=[]
         atom_site2_unp_accs=[]
-        #atom_site2_unp_names=[]                                                                                    
+        #atom_site2_unp_names=[]
         atom_site2_seq_nums=[]
         atom_site2_label_seq_ids=[]
         atom_site2_label_atom_ids=[]
@@ -191,12 +176,12 @@ class AnalysePisa:
             dist=round(float(distance),2)
             bond_type=bondtype
 
-            #*** Read uniprot accession and sequence numbers for atoms in bonds, from updated cif file                                                                       
+            #*** Read uniprot accession and sequence numbers for atoms in bonds, from updated cif file
 
             uniprot_info_1=self.read_uniprot_info(label_seqnum_1,seqnum_1,atname_1,res_1)
             unp_acc_1 = uniprot_info_1[0]
             unp_num_1 = uniprot_info_1[1]
-            
+
             uniprot_info_2=self.read_uniprot_info(label_seqnum_2,seqnum_2,atname_2,res_2)
             unp_acc_2 = uniprot_info_2[0]
             unp_num_2 = uniprot_info_2[1]
@@ -248,9 +233,9 @@ class AnalysePisa:
                       'atom_site_2_label_atom_ids': atom_site2_label_atom_ids,
                       'atom_site_2_inscodes': atom_site2_inscodes
                    }
-        
+
         return bond_dict
-            
+
     def get_molecules_dictionary(self,molecules):
         """
         Function creates molecules and residues dictionaries from xml data
@@ -266,10 +251,10 @@ class AnalysePisa:
         is_ligand: boolean type - is molecule class a 'Ligand'? 
         
         """
-        
+
         is_ligand=False
         molecules_dicts = []
-        
+
         for molecule in molecules:
             interface_residues_count = 0
             molecule_id = molecule.find('id').text
@@ -278,7 +263,7 @@ class AnalysePisa:
             #interface_molecules.append(molecule_class)
             residues_dicts=[]
 
-        
+
             if molecule_class in ['Ligand']:
                 is_ligand = True
             interface_residues = molecule.findall('residues/residue')
@@ -291,9 +276,9 @@ class AnalysePisa:
             solvation_energy_effects=[]
             residue_bonds=[]
 
-            
-            #Creating residues dictionaries                                                                                               
-            
+
+            #Creating residues dictionaries
+
             for residue in interface_residues:
                 residue_sernum=residue.find('ser_no').text
                 residue_name = residue.find('name').text
@@ -306,7 +291,7 @@ class AnalysePisa:
                 residue_bond= residue.find('bonds').text
 
                 #Writing interface residues dictionary
-                
+
                 residue_dict = { 'residue_sernum' : residue_sernum,
                                  'residue_name': residue_name,
                                  'residue_seqnum' : residue_seqnum,
@@ -327,9 +312,9 @@ class AnalysePisa:
                 buried_surface_areas.append(residue_bsa)
                 solvation_energy_effects.append(residue_solv_en)
 
-                #if str(residue_bsa) != "0":                                                                                                               
-                #residues_dicts.append(residue_dict)                                                                                                       
-                #if str(residue_bsa) != "0":                                                                                                               
+                #if str(residue_bsa) != "0":
+                #residues_dicts.append(residue_dict)
+                #if str(residue_bsa) != "0":
 
                 interface_residues_count += 1
 
@@ -354,9 +339,9 @@ class AnalysePisa:
             #if there is only one inteface residues, discard interface as valid interface
             if len(interface_residues) == 1:
                 is_ligand = True
-            
+
         return molecules_dicts,interface_residues_count,is_ligand
-    
+
     def process_pisa_interface_xml(self, interfaces_xml_file,assembly_xml_file):
         """
         Function writes assembly interfaces dictionaries
@@ -371,10 +356,10 @@ class AnalysePisa:
 
         """
         result = {}
-        
+
         if os.path.exists(interfaces_xml_file) and os.path.exists(assembly_xml_file):
-            asroot=self.parse_xml_file(xml_file=assembly_xml_file)
-            root = self.parse_xml_file(xml_file=interfaces_xml_file)
+            asroot=parse_xml_file(xml_file=assembly_xml_file)
+            root = parse_xml_file(xml_file=interfaces_xml_file)
             if root and asroot:
                 assembly_status=asroot.find('status').text
                 assemblies =asroot.findall('asu_complex')
@@ -391,11 +376,11 @@ class AnalysePisa:
                     assem_formula=assem.find('assembly/formula').text
                     assem_composition=assem.find('assembly/composition').text
 
-                
+
                 result['assembly_status']=assembly_status
 
                 #***** Round to two decimals some assembly properties ******
-                
+
                 assembly_mmsize = assem_mmsize
                 assembly_diss_energy=round(float(assem_diss_energy),2)
                 assembly_asa = round(float(assem_asa),2)
@@ -405,7 +390,7 @@ class AnalysePisa:
                 assembly_int_energy = round(float(assem_int_energy),2)
                 assembly_formula = assem_formula
                 assembly_composition = assem_composition
-                
+
                 #******** Create interfaces dictionaries **********************
                 status = root.find('status').text
                 num_interfaces = root.find('n_interfaces').text
@@ -417,7 +402,7 @@ class AnalysePisa:
 
                 non_ligand_interface_count = 0
 
-                
+
                 for interface in interfaces:
 
                     #** Interface General information
@@ -429,15 +414,15 @@ class AnalysePisa:
                     p_value = round(float(interface.find('pvalue').text),3)
 
                     #** No. of bonds counted
-                    
+
                     n_h_bonds = int(interface.find('h-bonds/n_bonds').text)
                     n_ss_bonds = int(interface.find('ss-bonds/n_bonds').text)
                     n_covalent_bonds = int(interface.find('cov-bonds/n_bonds').text)
                     n_salt_bridges = int(interface.find('salt-bridges/n_bonds').text)
                     other_contacts = int(interface.find('other-bonds/n_bonds').text)
-                    
+
                     #reading bonds
-                    
+
                     hbonds=interface.findall('h-bonds/bond')
                     sbridges=interface.findall('salt-bridges/bond')
                     covbonds=interface.findall('cov-bonds/bond')
@@ -451,21 +436,21 @@ class AnalysePisa:
                     ssbond_dict=self.get_bond_dictionary(ssbonds,'ss-bonds')
                     othbond_dict=self.get_bond_dictionary(othbonds,'other-bond')
 
-                    
+
                     #is_ligand = False
                     molecules_dicts = []
                     #interface_molecules = []
                     actual_interface_residues = []
                     molecules = interface.findall('molecule')
 
-                    
+
                     molecules_dicts,interface_residues_count,is_ligand=self.get_molecules_dictionary(molecules)
-                    
+
                     if not is_ligand:
                         non_ligand_interface_count += 1
 
                         interface_dict = {'interface_id': interface_id,
-                                      
+
                                       'interface_area': interface_area,
                                       'solvation_energy': interface_solvation_energy,
                                       'stabilization_energy': interface_stabilization_energy,
@@ -481,17 +466,17 @@ class AnalysePisa:
                                       'disulfide_bonds' : ssbond_dict,
                                       'covalent_bonds': covbond_dict,
                                       'other_bonds': othbond_dict,
-                                      'molecules' : molecules_dicts    
+                                      'molecules' : molecules_dicts
                                       }
 
                         #********** Append all dictionaries in 'result'  **********
-                        
+
                         result.setdefault('id', []).append(interface_id)
                         result.setdefault('int_area', []).append(interface_area)
                         result.setdefault('interface_dicts', []).append(interface_dict)
-                
+
                 #*********** Assembly information added to dictionary **********
-                
+
                 result['non_ligand_interface_count'] = non_ligand_interface_count
                 result['assembly_mmsize'] = assembly_mmsize
                 result['assembly_diss_energy'] = assembly_diss_energy
@@ -522,7 +507,7 @@ class AnalysePisa:
         """
         if interfaces_results:
             overall = len(interfaces_results.get('id', []))
-            
+
             interface_dicts = interfaces_results.get('interface_dicts', [])
 
             assem_dict={    'mmsize':interfaces_results.get('assembly_mmsize'),
@@ -546,7 +531,7 @@ class AnalysePisa:
                                  'assembly' : assem_dict
                                  }
             self.results.setdefault('PISA', assembly_dictionary)
-   
+
     def analyse_pisa_result(self, interfaces_xml_file, pdb_id, assembly_id,assembly_xml_file, entry_type='assembly'):
 
         """
@@ -578,20 +563,20 @@ class AnalysePisa:
         Returns: None
 
         """
-        
+
         pdbid_file = os.path.join(self.input_dir,self.input_cif_file) if self.input_cif_file else os.path.join(self.input_dir,'{}-assembly{}.cif.gz'.format(pdb_id,assembly_id))
         interfaces_xml_file = os.path.join(output_dir,'interfaces.xml')
         assembly_xml_file = os.path.join(output_dir, 'assembly.xml')
-               
+
         ok1= pdbid_file
-        
+
         if ok1:
             self.get_pisa_result(input_file=pdbid_file, interfaces_xml_file=interfaces_xml_file,
                                  assembly_xml_file=assembly_xml_file,pdb_id=pdb_id,pisa_config=pisa_config)
             self.analyse_pisa_result(interfaces_xml_file=interfaces_xml_file, pdb_id=pdb_id, assembly_id=assembly_id, entry_type='assembly',                                     assembly_xml_file=assembly_xml_file)
 
     def write_result_json(self):
-        
+
         if self.results:
             output_file = os.path.join(self.output_dir,self.result_json_file) if self.result_json_file else os.path.join(self.output_dir,'{}-assembly{}.json'.format(self.pdb_id,self.assembly_id))
             with open(output_file, 'w') as out_file:
